+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Support for package tagging: `list`, `add`, `clear`, `remove` and `replace` tags.
+
 ## [0.22.2] - 2020-06-11
 
 ### Added